#6111 feat(refactor): regulates use of react context
