--- src/components/SearchPane/Form/SearchPaneForm.test.tsx
@@ -1,12 +1,0 @@
-import React from 'react';
-import { shallow } from 'enzyme';
-
-import SearchPaneForm from './SearchPaneForm';
-
-describe('<SearchPaneForm />', () => {
-  const output = shallow(<SearchPaneForm />);
-
-  it('renders the component', () => {
-    expect(output);
-  });
-});
