+++ src/components/SearchPane/SearchPaneForm.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import Button from 'Button/Button';
+import Button from 'Button';
 
 export const SearchPaneForm = () => (
   <form
